serialization: provide deserialize-info from toplevel modules
This change makes it so that the following program runs successfully
after being converted to an executable using `raco exe`:

    #lang racket/base

    (require gregor
             racket/serialize)

    (deserialize (serialize (current-time)))
    (deserialize (serialize (today)))
    (deserialize (serialize (now/moment)))

As it stands, the deserialization process fails to look up the
`deserialization-info` submodules in an exe. On failure, that process
then has a fallback that looks at the toplevel module where the struct
is declared to get the info. So, this change fixes the issue by
providing the info directly from the toplevel modules in addition to
the `deserialization-info` submodules.

See racket/racket#4967 for more details.